`DeBruijnVar` takes a single type parameter
