+++ charon/src/ast/types/vars.rs
@@ -1,6 +1,6 @@
 //! Type-level variables. There are 4 kinds of variables at the type-level: regions, types, const
 //! generics and trait clauses. The relevant definitions are in this module.
-use derive_visitor::{Drive, DriveMut};
+use derive_visitor::{Drive, DriveMut, Event, Visitor, VisitorMut};
 use serde::{Deserialize, Serialize};
 
 use crate::ast::*;
@@ -48,12 +48,13 @@ pub struct DeBruijnId {
 ///                                   Var id: 1
 /// ```
 #[derive(Debug, PartialEq, Eq, Copy, Clone, Hash, PartialOrd, Ord, Serialize, Deserialize)]
-pub enum DeBruijnVar<Bound, Free> {
+// TODO: translate toplevel lifetimes as free vars
+pub enum DeBruijnVar<Id> {
     /// A variable attached to the nth binder, counting from the inside.
-    Bound(DeBruijnId, Bound),
+    Bound(DeBruijnId, Id),
     /// A variable attached to an implicit binder outside all other binders. This is not present in
     /// translated code, and only provided as a convenience for convenient variable manipulation.
-    Free(Free),
+    Free(Id),
 }
 
 // We need to manipulate a lot of indices for the types, variables, definitions, etc. In order not
@@ -112,10 +113,10 @@ pub struct TraitClause {
     pub trait_: PolyTraitDeclRef,
 }
 
-pub type RegionDbVar = DeBruijnVar<RegionId, RegionId>;
-pub type TypeDbVar = DeBruijnVar<TypeVarId, TypeVarId>;
-pub type ConstGenericDbVar = DeBruijnVar<ConstGenericVarId, ConstGenericVarId>;
-pub type ClauseDbVar = DeBruijnVar<TraitClauseId, TraitClauseId>;
+pub type RegionDbVar = DeBruijnVar<RegionId>;
+pub type TypeDbVar = DeBruijnVar<TypeVarId>;
+pub type ConstGenericDbVar = DeBruijnVar<ConstGenericVarId>;
+pub type ClauseDbVar = DeBruijnVar<TraitClauseId>;
 
 impl DeBruijnId {
     pub fn zero() -> Self {
@@ -143,20 +144,19 @@ impl DeBruijnId {
     }
 }
 
-impl<Bound, Free> DeBruijnVar<Bound, Free>
+impl<Id> DeBruijnVar<Id>
 where
-    Bound: Copy,
-    Free: Copy,
+    Id: Copy,
 {
-    pub fn new_at_zero(id: Bound) -> Self {
+    pub fn new_at_zero(id: Id) -> Self {
         DeBruijnVar::Bound(DeBruijnId::new(0), id)
     }
 
-    pub fn free(id: Free) -> Self {
+    pub fn free(id: Id) -> Self {
         DeBruijnVar::Free(id)
     }
 
-    pub fn bound(index: DeBruijnId, id: Bound) -> Self {
+    pub fn bound(index: DeBruijnId, id: Id) -> Self {
         DeBruijnVar::Bound(index, id)
     }
 
@@ -167,7 +167,7 @@ where
         }
     }
 
-    pub fn bound_at_depth(&self, depth: DeBruijnId) -> Option<Bound> {
+    pub fn bound_at_depth(&self, depth: DeBruijnId) -> Option<Id> {
         match *self {
             DeBruijnVar::Bound(dbid, varid) if dbid == depth => Some(varid),
             _ => None,
@@ -178,5 +178,38 @@ where
 impl TypeVar {
     pub fn new(index: TypeVarId, name: String) -> TypeVar {
         TypeVar { index, name }
+    }
+}
+
+// The derive macro doesn't handle generics.
+impl<Id: Drive> Drive for DeBruijnVar<Id> {
+    fn drive<V: Visitor>(&self, visitor: &mut V) {
+        visitor.visit(self, Event::Enter);
+        match self {
+            DeBruijnVar::Bound(x, y) => {
+                x.drive(visitor);
+                y.drive(visitor);
+            }
+            DeBruijnVar::Free(x) => {
+                x.drive(visitor);
+            }
+        }
+        visitor.visit(self, Event::Exit);
+    }
+}
+
+impl<Id: DriveMut> DriveMut for DeBruijnVar<Id> {
+    fn drive_mut<V: VisitorMut>(&mut self, visitor: &mut V) {
+        visitor.visit(self, Event::Enter);
+        match self {
+            DeBruijnVar::Bound(x, y) => {
+                x.drive_mut(visitor);
+                y.drive_mut(visitor);
+            }
+            DeBruijnVar::Free(x) => {
+                x.drive_mut(visitor);
+            }
+        }
+        visitor.visit(self, Event::Exit);
     }
 }
