+++ charon/src/bin/generate-ml/templates/Types.ml
@@ -67,18 +67,17 @@ class ['self] iter_const_generic_base_base =
     method visit_de_bruijn_id : 'env -> de_bruijn_id -> unit = fun _ _ -> ()
 
     method visit_de_bruijn_var
-        : 'b 'f.
-          ('env -> 'b -> unit) ->
-          ('env -> 'f -> unit) ->
+        : 'id.
+          ('env -> 'id -> unit) ->
           'env ->
-          ('b, 'f) de_bruijn_var ->
+          'id de_bruijn_var ->
           unit =
-      fun visit_bound visit_free env x ->
+      fun visit_id env x ->
         match x with
         | Bound (dbid, varid) ->
             self#visit_de_bruijn_id env dbid;
-            visit_bound env varid
-        | Free varid -> visit_free env varid
+            visit_id env varid
+        | Free varid -> visit_id env varid
   end
 
 (** Ancestor for map visitor for {!type: Types.ty} *)
@@ -90,20 +89,19 @@ class virtual ['self] map_const_generic_base_base =
       fun _ x -> x
 
     method visit_de_bruijn_var
-        : 'b 'f.
-          ('env -> 'b -> 'b) ->
-          ('env -> 'f -> 'f) ->
+        : 'id 'f.
+          ('env -> 'id -> 'id) ->
           'env ->
-          ('b, 'f) de_bruijn_var ->
-          ('b, 'f) de_bruijn_var =
-      fun visit_bound visit_free env x ->
+          'id de_bruijn_var ->
+          'id de_bruijn_var =
+      fun visit_id env x ->
         match x with
         | Bound (dbid, varid) ->
             let dbid = self#visit_de_bruijn_id env dbid in
-            let varid = visit_bound env varid in
+            let varid = visit_id env varid in
             Bound (dbid, varid)
         | Free varid ->
-            let varid = visit_free env varid in
+            let varid = visit_id env varid in
             Free varid
   end
 
@@ -115,20 +113,19 @@ class virtual ['self] reduce_const_generic_base_base =
       fun _ _ -> self#zero
 
     method visit_de_bruijn_var
-        : 'b 'f.
-          ('env -> 'b -> 'a) ->
-          ('env -> 'f -> 'a) ->
+        : 'id 'f.
+          ('env -> 'id -> 'a) ->
           'env ->
-          ('b, 'f) de_bruijn_var ->
+          'id de_bruijn_var ->
           'a =
-      fun visit_bound visit_free env x ->
+      fun visit_id env x ->
         match x with
         | Bound (dbid, varid) ->
             let acc1 = self#visit_de_bruijn_id env dbid in
-            let acc2 = visit_bound env varid in
+            let acc2 = visit_id env varid in
             self#plus acc1 acc2
         | Free varid ->
-            let acc = visit_free env varid in
+            let acc = visit_id env varid in
             acc
   end
 
@@ -140,20 +137,19 @@ class virtual ['self] mapreduce_const_generic_base_base =
       fun _ x -> (x, self#zero)
 
     method visit_de_bruijn_var
-        : 'b 'f.
-          ('env -> 'b -> 'b * 'a) ->
-          ('env -> 'f -> 'f * 'a) ->
+        : 'id 'f.
+          ('env -> 'id -> 'id * 'a) ->
           'env ->
-          ('b, 'f) de_bruijn_var ->
-          ('b, 'f) de_bruijn_var * 'a =
-      fun visit_bound visit_free env x ->
+          'id de_bruijn_var ->
+          'id de_bruijn_var * 'a =
+      fun visit_id env x ->
         match x with
         | Bound (dbid, varid) ->
             let dbid, acc1 = self#visit_de_bruijn_id env dbid in
-            let varid, acc2 = visit_bound env varid in
+            let varid, acc2 = visit_id env varid in
             (Bound (dbid, varid), self#plus acc1 acc2)
         | Free varid ->
-            let varid, acc = visit_free env varid in
+            let varid, acc = visit_id env varid in
             (Free varid, acc)
   end
 
@@ -279,10 +275,10 @@ type ('rid, 'id) g_region_group = {
 }
 [@@deriving show]
 
-type region_db_var = (region_id, region_id) de_bruijn_var [@@deriving show]
-type type_db_var = (type_var_id, type_var_id) de_bruijn_var [@@deriving show]
-type const_generic_db_var = (const_generic_var_id, const_generic_var_id) de_bruijn_var [@@deriving show]
-type trait_db_var = (trait_clause_id, trait_clause_id) de_bruijn_var [@@deriving show]
+type region_db_var = region_id de_bruijn_var [@@deriving show]
+type type_db_var = type_var_id de_bruijn_var [@@deriving show]
+type const_generic_db_var = const_generic_var_id de_bruijn_var [@@deriving show]
+type trait_db_var = trait_clause_id de_bruijn_var [@@deriving show]
 
 type region_var_group = (RegionId.id, RegionGroupId.id) g_region_group
 [@@deriving show]
